Half-Finished
Sadly can't finish it this week, because I won't be home and fully unavailable in the coming week.

This is not stable and does not properly safe data yet, but the rest should be somewhat working except for some of the locking and shops, which don't check for neighboring blocks yet.